Removed all mascot animations. Removed title from main xml. Added new images and icons for social fragment. changed navbar to be fragment
