--- app/src/main/res/drawable/boy_walk_animate.xml
@@ -1,12 +1,0 @@
-<animation-list xmlns:android="http://schemas.android.com/apk/res/android"
-    android:oneshot="false">
-    <item android:drawable="@drawable/boy_walk1" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk2" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk3" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk4" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk5" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk6" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk7" android:duration="200" />
-    <item android:drawable="@drawable/boy_walk8" android:duration="200" />
-
-</animation-list>
