--- app/src/main/res/drawable/girl_walk_animate.xml
@@ -1,12 +1,0 @@
-<animation-list xmlns:android="http://schemas.android.com/apk/res/android"
-    android:oneshot="false">
-    <item android:drawable="@drawable/girl_walk1" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk2" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk3" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk4" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk5" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk6" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk7" android:duration="200" />
-    <item android:drawable="@drawable/girl_walk8" android:duration="200" />
-
-</animation-list>
