+++ app/src/main/res/layout/activity_main.xml
@@ -22,7 +22,7 @@
             android:layout_height="wrap_content"
             >
 
-
+<!--
             <ImageView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -48,6 +48,7 @@
                 android:id="@+id/girl_walk_image"
                 android:layout_gravity="bottom|right"
                 />
+                -->
 
 
         </LinearLayout>
@@ -63,64 +64,13 @@
 
     </LinearLayout>
 
-    <android.support.percent.PercentRelativeLayout
-        android:id="@+id/InnerRelativeLayout"
+    <RelativeLayout
+        android:id="@+id/navbar_container"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@color/Black"
+        android:layout_height="100dp"
+        android:orientation="vertical"
         android:layout_alignParentBottom="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true">
+        />
 
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:srcCompat="@drawable/slider_bar"
-            android:id="@+id/highligherImage"
-            android:layout_alignLeft="@+id/SocialButton"
-            android:layout_alignRight="@+id/SocialButton"
-            android:layout_marginBottom="-2dp"
-            />
-
-
-        <Button
-            android:text="@string/Social"
-            app:layout_widthPercent="33%"
-            android:layout_height="wrap_content"
-            android:id="@+id/SocialButton"
-            android:theme="@style/IconButtonStyle"
-            android:drawableRight="@drawable/ic_social"
-            android:layout_alignParentLeft="true"
-            android:layout_below="@id/highligherImage"
-            />
-
-
-        <Button
-            android:text="@string/Learn"
-            app:layout_widthPercent="33%"
-            android:layout_height="wrap_content"
-            android:id="@+id/LearnButton"
-            android:theme="@style/IconButtonStyle"
-            android:drawableRight="@drawable/ic_learn"
-            android:layout_centerInParent="true"
-            android:layout_below="@id/highligherImage"
-            />
-
-
-
-        <Button
-            android:text="@string/Profile"
-            app:layout_widthPercent="33%"
-            android:layout_height="wrap_content"
-            android:id="@+id/ProfileButton"
-            android:layout_weight="1"
-            android:theme="@style/IconButtonStyle"
-            android:drawableRight="@drawable/ic_profile"
-            android:layout_alignParentRight="true"
-            android:layout_below="@id/highligherImage"
-            />
-
-
-    </android.support.percent.PercentRelativeLayout>
 
 </RelativeLayout>
