+++ app/src/main/res/layout/navbar_fragment.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/activity_main"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:background="#ffffff"
+    tools:context="com.mismatched.nowyouretalking.NavbarFragment"
+    android:orientation="vertical">
+
+
+<android.support.percent.PercentRelativeLayout
+    android:id="@+id/InnerRelativeLayout"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:background="@color/Black"
+    android:layout_alignParentBottom="true"
+    android:layout_alignParentLeft="true"
+    android:layout_alignParentStart="true">
+
+    <ImageView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:srcCompat="@drawable/slider_bar"
+        android:id="@+id/highligherImage"
+        android:layout_alignLeft="@+id/SocialButton"
+        android:layout_alignRight="@+id/SocialButton"
+        android:layout_marginBottom="-2dp"
+        />
+
+
+    <Button
+        android:text="@string/Social"
+        app:layout_widthPercent="33%"
+        android:layout_height="wrap_content"
+        android:id="@+id/SocialButton"
+        android:theme="@style/IconButtonStyle"
+        android:drawableRight="@drawable/ic_social"
+        android:layout_alignParentLeft="true"
+        android:layout_below="@id/highligherImage"
+        />
+
+
+    <Button
+        android:text="@string/Learn"
+        app:layout_widthPercent="33%"
+        android:layout_height="wrap_content"
+        android:id="@+id/LearnButton"
+        android:theme="@style/IconButtonStyle"
+        android:drawableRight="@drawable/ic_learn"
+        android:layout_centerInParent="true"
+        android:layout_below="@id/highligherImage"
+        />
+
+
+
+    <Button
+        android:text="@string/Profile"
+        app:layout_widthPercent="33%"
+        android:layout_height="wrap_content"
+        android:id="@+id/ProfileButton"
+        android:layout_weight="1"
+        android:theme="@style/IconButtonStyle"
+        android:drawableRight="@drawable/ic_profile"
+        android:layout_alignParentRight="true"
+        android:layout_below="@id/highligherImage"
+        />
+
+
+</android.support.percent.PercentRelativeLayout>
+    </RelativeLayout>
